Create stig_pkg_installs.sh
This file is used to install packages for meeting STIG requirements. 

Since SLE Micro is transactional, the install of a package is typically done with transactional-update pkg install <package_name> and a reboot. When you have multiple packages to install, you can use transactional-update shell to enter an interactive shell where you can run multiple things and it counts as a single transaction.